pyproject.toml
# pyproject.toml
[project]
name = "assistme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml",
    "typer",
    "playwright",
    "python-dotenv",
    "requests>=2.32.5",
]
